requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
cryptography==42.0.5
django-allauth==0.63.3
celery==5.3.6
redis==5.0.4
Pillow==10.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
